+++ src/main.ts
@@ -1,3 +1,3 @@
 export { Neatlint } from "./utils/Neatlint.util";
 
-export type { NeatlintOptions } from "./types/NeatlintOptions.type";
+export type * from "./barrels/Types.barrel"