+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Added:
   - **More rules.** Rules that prioritize type safety and logic errors.
+  - **NeatlintOptions.Disabled field.** ESLint completely frees up your settings.